Fix up division operators priority
One operator I wasn't sure of: 22C7 (divisiontimes). This is an oddball character that as far as I could see, has no common usage. I put it with the times chars in terms of priority rather than with division because it looks more like an asterisk than a division sign.

To get division sign in a good spot, I also moved "not sign" (00AC). It was has the wrong priority. It is higher than the other logical operators, but those are all pretty low (lower than relational operators). I grouped it with some other prefix operators that had the same spacing... not that I know whether the spacing is good or not...